Reorganize routing for Back-end
- Move User routing from app.jsx to userController.js
- Move Recipe routing from app.jsx to recipeController.js
