+++ back-end/app.jsx
@@ -6,13 +6,10 @@ const cookieParser = require("cookie-parser");
 // Replace the uri string with your MongoDB deployment's connection string.
 require("dotenv").config({ path: "./config.env" });
 const recipeController = require("./controllers/recipeController");
+const userController = require("./controllers/userController");
 const authController = require("./controllers/authController");
 const PORT = process.env.PORT || 5000;
 const URI = process.env.ATLAS_URI;
-
-// Import schema here!
-const UserModel = require("./models/Users");
-const Recipe = require("./models/Recipe");
 
 const app = express();
 //turns it to json file
@@ -31,77 +28,25 @@ mongoose
   .then(() => console.log("Connected to MongoDB..."))
   .catch((err) => console.log(err));
 
-//recipe routes
-app.get("/api/recipe/add", recipeController.addRecipe);
+// ================== Recipe routes ==================
 app.get("/api/recipe/all_recipes", recipeController.getAllRecipes);
 app.post("/uploadRecipe", recipeController.uploadRecipe);
+app.get(
+  "/api/recipe/all_recipes/:user_id",
+  recipeController.getRecipesbyUserID
+);
+app.get("/api/recipe/:id", recipeController.getRecipeByID);
 
+// ================== User routes ==================
+app.get("/api/:userid", userController.getProfileByID);
+app.get("/api/followings/:profileList", userController.getFollowings);
+app.post("/api/followProfile/:userID/:profileID", userController.followProfile);
+app.post(
+  "/api/unfollowProfile/:userID/:profileID",
+  userController.unfollowProfile
+);
 
-app.get("/api/recipe/all_recipes/:user_id", (req, res) => {
-  Recipe.find(
-    { author_id: req.params.user_id },
-    {
-      title: 1,
-      cover_image: 1,
-      _id: 1,
-    }
-  )
-    .then((result) => {
-      res.send(result);
-    })
-    .catch((err) => console.log(err));
-});
-
-app.get("/api/recipe/:id", (req, res) => {
-  Recipe.findById(req.params.id)
-    .then((result) => {
-      res.send(result);
-    })
-    .catch((err) => console.log(err));
-});
-
-app.get("/api/:userid", (req, res) => {
-  UserModel.find({ _id: req.params.userid }, { password: 0 })
-    .then((result) => {
-      res.send(result);
-    })
-    .catch((err) => console.log(err));
-});
-
-app.get("/api/followings/:profileList", (req, res) => {
-  const profileList = req.params.profileList.split(",");
-  UserModel.find({ _id: { $in: profileList } }, { password: 0 })
-    .then((result) => {
-      res.send(result);
-    })
-    .catch((err) => console.log(err));
-});
-
-app.post("/api/followProfile/:userID/:profileID", (req, res) => {
-  UserModel.findByIdAndUpdate(
-    req.params.userID,
-    { $push: { followings: req.params.profileID } },
-    { new: true }
-  )
-    .then((result) => {
-      res.send(result);
-    })
-    .catch((err) => console.log(err));
-});
-
-app.post("/api/unfollowProfile/:userID/:profileID", (req, res) => {
-  UserModel.findByIdAndUpdate(
-    req.params.userID,
-    { $pull: { followings: req.params.profileID } },
-    { new: true }
-  )
-    .then((result) => {
-      res.send(result);
-    })
-    .catch((err) => console.log(err));
-});
-
-//authentication routes
+// =============== authentication routes ===========
 app.post("/login", authController.login);
 app.post("/logout", authController.logout);
 app.post("/signup", authController.signup);
